Conveyor logger improvements
