--- ledger-core/v2/virtual/statemachine/logger.go
@@ -1,250 +1,0 @@
-// Copyright 2020 Insolar Network Ltd.
-// All rights reserved.
-// This material is licensed under the Insolar License version 1.0,
-// available at https://github.com/insolar/assured-ledger/blob/master/LICENSE.md.
-
-package statemachine
-
-import (
-	"context"
-	"fmt"
-	"reflect"
-	"runtime"
-	"strings"
-
-	"github.com/insolar/assured-ledger/ledger-core/v2/conveyor"
-	"github.com/insolar/assured-ledger/ledger-core/v2/conveyor/smachine"
-	"github.com/insolar/assured-ledger/ledger-core/v2/instrumentation/inslogger"
-	"github.com/insolar/assured-ledger/ledger-core/v2/log"
-	"github.com/insolar/assured-ledger/ledger-core/v2/log/global"
-	"github.com/insolar/assured-ledger/ledger-core/v2/log/logfmt"
-	"github.com/insolar/assured-ledger/ledger-core/v2/vanilla/throw"
-	"github.com/insolar/assured-ledger/ledger-core/v2/virtual/integration/convlog"
-)
-
-var _ smachine.StepLogger = ConveyorLogger{}
-
-type ConveyorLogger struct {
-	tracerCtx context.Context
-	tracerID  smachine.TracerID
-
-	logger log.Logger
-}
-
-func (c ConveyorLogger) LogInternal(data smachine.StepLoggerData, updateType string) {
-	// TODO LogInternal
-}
-
-func (c ConveyorLogger) LogAdapter(data smachine.StepLoggerData, adapterID smachine.AdapterID, callID uint64, fields []logfmt.LogFieldMarshaller) {
-	// TODO LogAdapter
-}
-
-func (c ConveyorLogger) GetTracerID() smachine.TracerID {
-	return c.tracerID
-}
-
-func (c ConveyorLogger) GetLoggerContext() context.Context {
-	return c.tracerCtx
-}
-
-func (c ConveyorLogger) CanLogEvent(eventType smachine.StepLoggerEvent, stepLevel smachine.StepLogLevel) bool {
-	return true
-}
-
-func (c ConveyorLogger) CreateAsyncLogger(ctx context.Context, _ *smachine.StepLoggerData) (context.Context, smachine.StepLogger) {
-	return c.tracerCtx, c
-}
-
-type LogStepMessage struct {
-	*log.Msg
-
-	Message   string
-	Component string `txt:"sm"`
-	TraceID   string `opt:""`
-
-	MachineName interface{} `fmt:"%T"`
-	MachineID   string
-	SlotID      smachine.SlotID
-	SlotStepNo  uint32
-	CycleNo     uint32
-	From        string
-	To          string `opt:""`
-
-	Error     string `opt:""`
-	Backtrace string `opt:""`
-
-	ExecutionTime  int64 `opt:""`
-	InactivityTime int64 `opt:""`
-}
-
-func getStepName(step interface{}) string {
-	fullName := runtime.FuncForPC(reflect.ValueOf(step).Pointer()).Name()
-	if lastIndex := strings.LastIndex(fullName, "."); lastIndex >= 0 {
-		fullName = fullName[lastIndex+1:]
-	}
-	if lastIndex := strings.LastIndex(fullName, "-"); lastIndex >= 0 {
-		fullName = fullName[:lastIndex]
-	}
-
-	return fullName
-}
-
-func prepareStepName(sd *smachine.StepDeclaration) {
-	if !sd.IsNameless() {
-		return
-	}
-	sd.Name = getStepName(sd.Transition)
-}
-
-func (c ConveyorLogger) LogEvent(data smachine.StepLoggerData, msg interface{}, fields []logfmt.LogFieldMarshaller) {
-	c.logger.Errorm(msg, fields...)
-}
-
-func (c ConveyorLogger) LogUpdate(stepLoggerData smachine.StepLoggerData, stepLoggerUpdateData smachine.StepLoggerUpdateData) {
-	special := ""
-
-	switch stepLoggerData.EventType {
-	case smachine.StepLoggerUpdate:
-	case smachine.StepLoggerMigrate:
-		special = "migrate "
-	default:
-		panic(throw.Impossible())
-	}
-
-	prepareStepName(&stepLoggerData.CurrentStep)
-	prepareStepName(&stepLoggerUpdateData.NextStep)
-
-	suffix := ""
-	if stepLoggerData.Flags&smachine.StepLoggerDetached != 0 {
-		suffix = " (detached)"
-	}
-
-	if _, ok := stepLoggerData.Declaration.(*conveyor.PulseSlotMachine); ok {
-		return
-	}
-
-	var (
-		backtrace string
-		err       string
-	)
-	if stepLoggerData.Error != nil {
-		var st throw.StackTrace
-		err, st = throw.ErrorWithTopOrMinimizedStack(stepLoggerData.Error, convlog.StackMinimizePackage, true)
-		if st != nil {
-			backtrace = st.StackTraceAsText()
-		}
-	}
-
-	if err != "" {
-		fmt.Println(err)
-	}
-
-	msg := LogStepMessage{
-		Message: special + stepLoggerUpdateData.UpdateType + suffix,
-
-		MachineName: stepLoggerData.Declaration,
-		MachineID:   stepLoggerData.StepNo.MachineID(),
-		CycleNo:     stepLoggerData.CycleNo,
-		SlotID:      stepLoggerData.StepNo.SlotID(),
-		SlotStepNo:  stepLoggerData.StepNo.StepNo(),
-
-		From: stepLoggerData.CurrentStep.GetStepName(),
-		To:   stepLoggerUpdateData.NextStep.GetStepName(),
-
-		Error:     err,
-		Backtrace: backtrace,
-	}
-
-	if stepLoggerUpdateData.ActivityNano > 0 {
-		msg.ExecutionTime = stepLoggerUpdateData.ActivityNano.Nanoseconds()
-	}
-	if stepLoggerUpdateData.InactivityNano > 0 {
-		msg.InactivityTime = stepLoggerUpdateData.InactivityNano.Nanoseconds()
-	}
-
-	if stepLoggerData.Error == nil {
-		c.logger.Debug(msg)
-	} else {
-		c.logger.Error(msg)
-	}
-}
-
-type ConveyorLoggerFactory struct {
-}
-
-func (c ConveyorLoggerFactory) CreateStepLogger(ctx context.Context, _ smachine.StateMachine, traceID smachine.TracerID) smachine.StepLogger {
-	ctxWithTrace, logger := inslogger.WithTraceField(ctx, traceID)
-	return &ConveyorLogger{
-		tracerCtx: ctxWithTrace,
-		tracerID:  traceID,
-		logger:    logger,
-	}
-}
-
-type LogInternal struct {
-	*log.Msg `txt:"internal"`
-
-	Message   string `fmt:"internal - %s"`
-	Component string `txt:"sm"`
-
-	MachineID  string
-	SlotID     smachine.SlotID
-	SlotStepNo uint32
-	CycleNo    uint32
-	Error      error  `opt:""`
-	Backtrace  string `opt:""`
-}
-
-func (ConveyorLoggerFactory) LogMachineInternal(slotMachineData smachine.SlotMachineData, msg string) {
-	backtrace := ""
-	if slotMachineData.Error != nil {
-		if st := throw.DeepestStackTraceOf(slotMachineData.Error); st != nil {
-			st = throw.MinimizeStackTrace(st, convlog.StackMinimizePackage, true)
-			backtrace = st.StackTraceAsText()
-		}
-	}
-	global.Logger().Error(LogInternal{
-		Message: msg,
-
-		MachineID:  slotMachineData.StepNo.MachineID(),
-		CycleNo:    slotMachineData.CycleNo,
-		SlotID:     slotMachineData.StepNo.SlotID(),
-		SlotStepNo: slotMachineData.StepNo.StepNo(),
-		Error:      slotMachineData.Error,
-		Backtrace:  backtrace,
-	})
-}
-
-type LogCritical struct {
-	*log.Msg `txt:"internal"`
-
-	Message   string `fmt:"internal critical - %s"`
-	Component string `txt:"sm"`
-
-	MachineID  string
-	SlotID     smachine.SlotID
-	SlotStepNo uint32
-	CycleNo    uint32
-	Error      error  `opt:""`
-	Backtrace  string `opt:""`
-}
-
-func (ConveyorLoggerFactory) LogMachineCritical(slotMachineData smachine.SlotMachineData, msg string) {
-	backtrace := ""
-	if slotMachineData.Error != nil {
-		if st := throw.DeepestStackTraceOf(slotMachineData.Error); st != nil {
-			st = throw.MinimizeStackTrace(st, convlog.StackMinimizePackage, true)
-			backtrace = st.StackTraceAsText()
-		}
-	}
-	global.Logger().Error(LogCritical{
-		Message: msg,
-
-		MachineID:  slotMachineData.StepNo.MachineID(),
-		CycleNo:    slotMachineData.CycleNo,
-		SlotID:     slotMachineData.StepNo.SlotID(),
-		SlotStepNo: slotMachineData.StepNo.StepNo(),
-		Error:      slotMachineData.Error,
-		Backtrace:  backtrace,
-	})
-}
